+++ apps/tumeet/src/app/api/meet-together/plans/[planId]/route.ts
@@ -1,70 +1,3 @@
-import { createAdminClient } from '@tuturuuu/supabase/next/server';
-import { parseTimeFromTimetz } from '@tuturuuu/utils/time-helper';
-import { NextResponse } from 'next/server';
+import { DELETE, PATCH } from '@tuturuuu/apis/tumeet/plans/planId/route';
 
-interface Params {
-  params: Promise<{
-    planId: string;
-  }>;
-}
-
-// Utility function to parse time from timetz format (e.g., "09:00:00+00")
-
-export async function PUT(req: Request, { params }: Params) {
-  const sbAdmin = await createAdminClient();
-  const { planId: id } = await params;
-
-  const data = await req.json();
-
-  // Backend validation: ensure end_time is after start_time
-  if (data.start_time && data.end_time) {
-    const startHour = parseTimeFromTimetz(data.start_time);
-    const endHour = parseTimeFromTimetz(data.end_time);
-
-    if (
-      startHour !== undefined &&
-      endHour !== undefined &&
-      endHour <= startHour
-    ) {
-      return NextResponse.json(
-        { message: 'End time must be after start time' },
-        { status: 400 }
-      );
-    }
-  }
-
-  const { error } = await sbAdmin
-    .from('meet_together_plans')
-    .update(data)
-    .eq('id', id);
-
-  if (error) {
-    console.log(error);
-    return NextResponse.json(
-      { message: 'Error updating meet together plan' },
-      { status: 500 }
-    );
-  }
-
-  return NextResponse.json({ message: 'success' });
-}
-
-export async function DELETE(_: Request, { params }: Params) {
-  const sbAdmin = await createAdminClient();
-  const { planId: id } = await params;
-
-  const { error } = await sbAdmin
-    .from('meet_together_plans')
-    .delete()
-    .eq('id', id);
-
-  if (error) {
-    console.log(error);
-    return NextResponse.json(
-      { message: 'Error deleting meet together plan' },
-      { status: 500 }
-    );
-  }
-
-  return NextResponse.json({ message: 'success' });
-}
+export { PATCH, DELETE };
