+++ packages/apis/biome.json
@@ -1,0 +1,5 @@
+{
+  "$schema": "https://biomejs.dev/schemas/2.1.3/schema.json",
+  "root": false,
+  "extends": "//"
+}
