+++ packages/apis/eslint.config.mjs
@@ -1,0 +1,4 @@
+import { config } from '@tuturuuu/eslint-config/react-internal.js';
+
+/** @type {import("eslint").Linter.Config} */
+export default config;
