+++ packages/apis/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@tuturuuu/apis",
+  "license": "MIT",
+  "version": "0.0.10",
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "lint": "eslint .",
+    "test": "vitest"
+  },
+  "dependencies": {
+    "@tuturuuu/supabase": "workspace:*",
+    "@tuturuuu/utils": "workspace:*",
+    "dayjs": "^1.11.13",
+    "next": "^15.4.5",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "zod": "^4.0.14"
+  },
+  "devDependencies": {
+    "@tuturuuu/eslint-config": "workspace:*",
+    "@tuturuuu/types": "workspace:*",
+    "@tuturuuu/typescript-config": "workspace:*",
+    "@types/node": "^24.2.0",
+    "@types/react": "^19.1.9",
+    "typescript": "^5.9.2"
+  },
+  "exports": {
+    "./*": "./src/*.ts"
+  },
+  "packageManager": "bun@1.2.19"
+}
